Rename all composable file to match what they contain
instead of this weird kebab case, and add the "use" keyword before each file (matching the name of the function)
