+++ src/modules/map/components/openlayers/OpenLayersHighlightedFeatures.vue
@@ -17,8 +17,8 @@ import FeatureEdit from '@/modules/infobox/components/FeatureEdit.vue'
 import FeatureList from '@/modules/infobox/components/FeatureList.vue'
 import { useLayerZIndexCalculation } from '@/modules/map/components/common/z-index.composable'
 import OpenLayersPopover from '@/modules/map/components/openlayers/OpenLayersPopover.vue'
-import useVectorLayer from '@/modules/map/components/openlayers/utils/add-vector-layer-to-map.composable'
 import { highlightFeatureStyle } from '@/modules/map/components/openlayers/utils/markerStyle'
+import useVectorLayer from '@/modules/map/components/openlayers/utils/useVectorLayer.composable'
 import { FeatureInfoPositions } from '@/store/modules/ui.store'
 import { WGS84 } from '@/utils/coordinates/coordinateSystems'
 import { transformIntoTurfEquivalent } from '@/utils/geoJsonUtils'
