+++ src/modules/map/components/openlayers/OpenLayersKMLLayer.vue
@@ -8,7 +8,7 @@ import { useStore } from 'vuex'
 
 import KMLLayer from '@/api/layers/KMLLayer.class'
 import { IS_TESTING_WITH_CYPRESS } from '@/config'
-import useAddLayerToMap from '@/modules/map/components/openlayers/utils/add-layers-to-map.composable'
+import useAddLayerToMap from '@/modules/map/components/openlayers/utils/useAddLayerToMap.composable'
 import { parseKml } from '@/utils/kmlUtils'
 import log from '@/utils/logging'
