+++ src/modules/map/components/openlayers/OpenLayersVectorLayer.vue
@@ -15,7 +15,7 @@ import { computed, inject, toRefs, watch } from 'vue'
 
 import GeoAdminVectorLayer from '@/api/layers/GeoAdminVectorLayer.class'
 import { VECTOR_TILES_IMAGERY_STYLE_ID } from '@/config'
-import useAddLayerToMap from '@/modules/map/components/openlayers/utils/add-layers-to-map.composable'
+import useAddLayerToMap from '@/modules/map/components/openlayers/utils/useAddLayerToMap.composable'
 import log from '@/utils/logging'
 
 const props = defineProps({
