+++ src/modules/map/components/openlayers/utils/useAddLayerToMap.composable.js
