+++ src/modules/map/components/openlayers/utils/useMapInteractions.composable.js
