+++ src/modules/map/components/openlayers/utils/usePrintAreaRenderer.composable.js
