+++ src/modules/map/components/openlayers/utils/useViewBasedOnProjection.composable.js
